+++ CHANGELOG.rst
@@ -10,47 +10,112 @@ Changelog
 
 6.0.0 (Unreleased)
 ==================
-- Python 3.7 and below no longer supported.
+- 🚨 Python 3.7 and below no longer supported.
 - All models are now Pydantic models, for better data validation, serialization and deserialization.
 - Models no longer contain methods to parse content, these have been moved to dedicated parser classes.
-- The location of models inside the module changed, so all imports need to be modified.
+- 🚨 The location of models inside the module changed, so all imports need to be modified.
 - Model changes:
-    - Removed ``world`` from ``OnlineCharacter``.
-    - Removed ``category_icon`` from ``NewsEntry`` and ``News``.
-    - Added ``big_icon_url`` and ``small_icon_url`` to ``NewsCategory``
-    - ``Leaderboard.page`` renamed to ``current_page``.
-    - ``Highscores.page`` renamed to ``current_page``.
-    - ``Highscores.last_updated`` is now a ``datetime`` instead of a ``timedelta``.
-    - Removed ``owner``, ``world``, ``status`` and ``type`` from ``GuildHouse``.
-    - Removed ``name`` from ``Death``.
-    - Removed ``owner``, ``status`` and ``type`` from ``CharacterHouse``.
+    - ``Character``:
+        - ``account_status`` field replaced with ``is_premium``.
+        - ``traded`` field renamed to ``is_traded``.
+        - ``deleted`` property renamed to ``is_scheduled_for_deletion``.
+        - ``hidden`` property renamed to ``is_hidden``.
+    - ``Achievement``:
+        - ``secret`` field renamed to ``is_secret``
+    - ``CharacterHouse``:
+        - ``owner`` field removed.
+        - ``status`` field removed.
+        - ``type`` field removed.
+        - ``paid_until_date`` field renamed tp ``paid_until``.
+        - Removed ``owner``, ``status`` and ``type`` from ``CharacterHouse``.
+    - ``Killer``:
+        - Model name renamed to ``DeathParticipant``.
+        - ``player`` field renamed to ``is_player``.
+        - ``traded`` field renamed to ``is_traded``.
+    - ``Death``:
+        - ``by_player`` property renamed to ``is_by_player``
+        - ``name`` field removed.
+    - ``OtherCharacter``:
+        - ``online`` field renamed to ``is_online``.
+        - ``deleted`` field renamed to ``is_deleted``.
+        - ``traded`` field renamed to ``is_traded``.
+        - ``main`` field renamed to ``is_main``.
+    - ``WorldEntry`` / ``World``:
+        - ``online`` field renamed to ``is_online``
+        - ``battleye_date`` field renamed to ``battleye_since``
+        - ``premium_only`` field renamed to ``is_premium_only``
+        - ``battleye_protected`` property renamed to ``is_battleye_protected``
+    - ``OnlineCharacter``:
+        - ``world`` field removed.
+    - ``SpellsSection``:
+        - ``premium`` field renamed to ``is_premium``
+    - ``SpellEntry``/``Spell``:
+        - ``premium`` field renamed to ``is_premium``
+    - ``News``/``NewsEntry``:
+        - ``category_icon`` field removed. Can be accessed through ``category.big_icon_url`` and ``category.small_icon_url``.
+        - ``date`` field renamed to ``published_on``
+    - ``NewsArchive``:
+        - ``start_date`` field renamed to ``from_date``.
+        - ``end_date`` field renamed to ``to_date``.
+        - ``types`` field is now of type ``set``.
+        - ``categories`` field is now of type ``set``.
+    - ``NewsCategory`` (enum)
+        - Added ``big_icon_url`` and ``small_icon_url`` properties.
+    - ``Leaderboard``:
+        - ``current`` field renamed to ``is_current``.
+        - ``page`` field renamed to ``current_page``.
+    - ``Highscores``:
+        - ``.page`` field renamed to ``current_page``.
+        - ``last_updated`` field is now a ``datetime`` instead of a ``timedelta``.
+    - ``GuildMember``:
+        - ``joined`` field renamed to ``joined_on``.
+        - ``online`` field renamed to ``is_online``.
+    - ``GuildInvite``:
+        - ``date`` field renamed to ``invited_on``.
+    - ``GuildHouse``:
+        - ``paid_until_date`` field renamed tp ``paid_until``.
+        - ``owner`` field removed.
+        - ``world`` field removed.
+        - ``status`` field removed.
+        - ``type`` field removed.
+    - ``ForumBoard``:
+        - ``threads`` field renamed to ``entries``.
+        - ``page`` field renamed to ``current_page``.
+        - ``section_id`` field added.
+        - ``results_count`` field added.
+    - ``ForumThread``:
+        - ``posts`` field renamed to ``entries``.
+        - ``board_id`` field added.
+        - ``section_id`` field added.
+        - ``results_count`` field added.
+        - ``previous_topic_number`` may be ``None`` instead of ``0``.
+        - ``next_topic_number`` may be ``None`` instead of ``0``.
+    - ``LastPost``.
+        - ``date`` field renamed to ``posted_on``.
+        - ``deleted`` field renamed to ``is_author_deleted``.
+        - ``traded`` field renamed to ``is_author_traded``.
+    - ``CMPostArchive``:
+        - ``start_date`` field renamed to ``from_date``
+        - ``end_date`` field renamed to ``to_date``
+        - ``page`` field renamed to ``current_page``
+        - ``posts`` field renamed to ``entries``
+    - ``CMPost``:
+        - ``date`` field renamed to ``posted_on``
     - ``Auction`` renamed to ``AuctionDetails`` and is no longer a subclass of ``AuctionEntry``.
     - ``AuctionEntry`` renamed to ``Auction``.
-    - ``CharacterBazaar.page`` renamed to ``current_page``.
-    - ``NewsArchive.categories`` and ``NewsArchive.types`` are now sets instead of lists.
-    - ``Character.account_status`` replaced with ``is_premium``.
+    - ``CharacterBazaar``:
+        - ``page`` field renamed to ``current_page``.
     - Renamed ``DisplayImage`` to ``ItemEntry``.
     - Renamed ``DisplayMount`` to ``MountEntry``.
     - Renamed ``DisplayOutfit`` to ``OutfitEntry``.
     - Renamed ``DisplayFamiliar`` to ``FamiliarEntry``.
     - Renamed ``page`` to ``current_page``, ``results`` to ``results_count``, and ``fully_fetched`` to ``is_fully_fetched`` in ``ItemSummary``, ``Mounts``, ``Familiars`` and ``Outfits``.
-    - Renamed ``CMPostArchive`` attributes:
-        - ``start_date`` to ``from_date``
-        - ``end_date`` to ``to_date``
-        - ``page`` to ``current_page``
-        - ``posts`` to ``entries``
-    - Renamed ``CMPost.date``to ``posted_on``
-    - Renamed ``Killer``to ``DeathParticipant``
-    - Renamed ``ForumBoard.threads`` to ``entries``.
-    - Renamed ``ForumThread.posts`` to ``entries``.
-    - Renamed ``GuildInvite.date`` to ``invited_on``.
-    - Renamed ``LastPost.date`` to ``posted_on``.
+
+
+
     - Added ``thread_starter_deleted`` to ``ThreadEntry``.
-    - ``ForumThread.previous_topic_number`` and ``ForumThread.next_topic_number`` may be ``None`` instead of ``0``.
-    - Renamed ``News`` attributes:
-        - ``start_date`` to ``from_date``
-        - ``end_date`` to ``to_date``
-        - ``date`` to ``published_on``
+
+
 - Renamed ``Category`` to ``HighscoresCategory``.
 - Renamed ``BattlEyeTypeFilter`` to ``AuctionBattlEyeFilter``.
 - Renamed ``VocationFilter`` to ``HighscoresProfession``.